* libbase/utility.h, 	  server/types.h:  make sure truncation is used when convert float to integer. 	* server/matrix.{h,cpp} : use double instead of float for ActionScripts, use 	  integers for x_translation and y_translation. Add a template function transform(point2d), 	  since renderer agg and ogl use different point formats. 	* server/character.cpp, server/sprite_instance.cpp: take care of data types, minor cleanups.

+++ server/types.h
@@ -12,7 +12,8 @@
 #include <boost/cstdint.hpp> // for boost::?int??_t 
 
 inline double TWIPS_TO_PIXELS(int i) { return i / 20.0; }
-inline int PIXELS_TO_TWIPS(double d) { return d * 20; }
+// truncate when overflow occurs.
+inline int PIXELS_TO_TWIPS(double d) { return (int)(unsigned int)(d * 20); }
 
 namespace gnash {
 	class SWFStream;	// forward declaration
